Unify allPossibleTurns
